Create macro to implement ServicePreferences
