+++ core/engine/src/builtins/intl/number_format/options.rs
@@ -9,7 +9,7 @@ use boa_macros::js_str;
 use icu_decimal::{
     DecimalFormatterPreferences, preferences::NumberingSystem, provider::DecimalSymbolsV1,
 };
-use icu_locale::{extensions::unicode::Value, preferences::PreferenceKey};
+use icu_locale::extensions::unicode::Value;
 use icu_provider::{
     DataMarkerAttributes,
     prelude::icu_locale_core::{
@@ -1275,31 +1275,5 @@ impl ServicePreferences for DecimalFormatterPreferences {
         });
     }
 
-    fn as_unicode(&self) -> unicode::Unicode {
-        let mut exts = unicode::Unicode::new();
-
-        if let Some(nu) = self.numbering_system
-            && let Some(value) = nu.unicode_extension_value()
-        {
-            exts.keywords.set(unicode::key!("nu"), value);
-        }
-        exts
-    }
-
-    fn extended(&self, other: &Self) -> Self {
-        let mut result = *self;
-        result.extend(*other);
-        result
-    }
-
-    fn intersection(&self, other: &Self) -> Self {
-        let mut inter = *self;
-        if inter.locale_preferences != other.locale_preferences {
-            inter.locale_preferences = LocalePreferences::default();
-        }
-        if inter.numbering_system != other.numbering_system {
-            inter.numbering_system.take();
-        }
-        inter
-    }
+    impl_service_preferences!(numbering_system);
 }
